Validation on Enter key made optional.
validate method doesn't make the validation state NOT_VALIDATED prior to
validation. NOT_VALIDATED state is set only when text of the field
changes. This avoid to trigger a VALID-NOT_VALIDATED-VALID chain every
time the validate() method is called which was in practice triggering
the state change listeners everytime the validate() was called, even if
the state wasn't actually changed.