close #33: default values in split by size
Added a couple of values to the size combo. It's currently an editable
combo where the user can input integer number of bytes, KB or MB. This
should be enough to cover every need, also considering that "split by
size" is an imprecise task and it split roughly at the given size. To
solve the suggested use case simply input 1500 KB